+++ examples/apps/react-todomvc-ts/src/store/index.ts
@@ -12,4 +12,5 @@ export const store = createStore<TodoListEntity, TodoListState>({
   middlewares,
 })
 
-export const { Provider, useSelector, useNotify } = createReactStore(store)
+export const { Provider, useSelector, useEntity, useNotify } =
+  createReactStore(store)